move MI construction test
